refactor: extract PDF semantic recovery
